+++ lib/config.sh
@@ -9,12 +9,14 @@ declare -gA UL2CP_LOGIN_LOOP=()
 declare -gA UL2CP_LOGIN_MATCH=()
 declare -gA UL2CP_LOGIN_LOG=()
 declare -gA UL2CP_LOGIN_PORT_HINT=()
+declare -gA UL2CP_LOGIN_READY_MATCH=()
 declare -gA UL2CP_GAME_ENABLED=()
 declare -gA UL2CP_GAME_DIR=()
 declare -gA UL2CP_GAME_LOOP=()
 declare -gA UL2CP_GAME_MATCH=()
 declare -gA UL2CP_GAME_LOG=()
 declare -gA UL2CP_GAME_PORT_HINT=()
+declare -gA UL2CP_GAME_READY_MATCH=()
 
 reset_server_vars() {
   SERVER_ID=""
@@ -27,6 +29,7 @@ reset_server_vars() {
   LOGIN_MATCH="AuthServer"
   LOGIN_LOG="log/stdout.log"
   LOGIN_PORT_HINT=""
+  LOGIN_READY_MATCH=""
 
   GAME_ENABLED="false"
   GAME_DIR=""
@@ -34,6 +37,7 @@ reset_server_vars() {
   GAME_MATCH="GameServer"
   GAME_LOG="log/stdout.log"
   GAME_PORT_HINT=""
+  GAME_READY_MATCH=""
 }
 
 load_server_file() {
@@ -57,6 +61,7 @@ load_server_file() {
   UL2CP_LOGIN_MATCH["$SERVER_ID"]="$LOGIN_MATCH"
   UL2CP_LOGIN_LOG["$SERVER_ID"]="$LOGIN_LOG"
   UL2CP_LOGIN_PORT_HINT["$SERVER_ID"]="$LOGIN_PORT_HINT"
+  UL2CP_LOGIN_READY_MATCH["$SERVER_ID"]="$LOGIN_READY_MATCH"
 
   UL2CP_GAME_ENABLED["$SERVER_ID"]="$GAME_ENABLED"
   UL2CP_GAME_DIR["$SERVER_ID"]="$GAME_DIR"
@@ -64,6 +69,7 @@ load_server_file() {
   UL2CP_GAME_MATCH["$SERVER_ID"]="$GAME_MATCH"
   UL2CP_GAME_LOG["$SERVER_ID"]="$GAME_LOG"
   UL2CP_GAME_PORT_HINT["$SERVER_ID"]="$GAME_PORT_HINT"
+  UL2CP_GAME_READY_MATCH["$SERVER_ID"]="$GAME_READY_MATCH"
 }
 
 load_server_configs() {
@@ -71,6 +77,22 @@ load_server_configs() {
   local found
 
   UL2CP_SERVER_IDS=()
+  UL2CP_SERVER_NAMES=()
+  UL2CP_SERVER_OWNERS=()
+  UL2CP_LOGIN_ENABLED=()
+  UL2CP_LOGIN_DIR=()
+  UL2CP_LOGIN_LOOP=()
+  UL2CP_LOGIN_MATCH=()
+  UL2CP_LOGIN_LOG=()
+  UL2CP_LOGIN_PORT_HINT=()
+  UL2CP_LOGIN_READY_MATCH=()
+  UL2CP_GAME_ENABLED=()
+  UL2CP_GAME_DIR=()
+  UL2CP_GAME_LOOP=()
+  UL2CP_GAME_MATCH=()
+  UL2CP_GAME_LOG=()
+  UL2CP_GAME_PORT_HINT=()
+  UL2CP_GAME_READY_MATCH=()
   found=0
   shopt -s nullglob
   for file in "$UNIX_L2_CP_SERVER_DIR"/*.conf; do
@@ -145,6 +167,14 @@ component_port_hint() {
   case "$2" in
     login) printf '%s\n' "${UL2CP_LOGIN_PORT_HINT[$1]:-}" ;;
     game) printf '%s\n' "${UL2CP_GAME_PORT_HINT[$1]:-}" ;;
+    *) return 1 ;;
+  esac
+}
+
+component_ready_match() {
+  case "$2" in
+    login) printf '%s\n' "${UL2CP_LOGIN_READY_MATCH[$1]:-}" ;;
+    game) printf '%s\n' "${UL2CP_GAME_READY_MATCH[$1]:-}" ;;
     *) return 1 ;;
   esac
 }